Include the solution for the unisex bathroom problem, using monitors
