+++ 02 - Synchronization in concurrent programs/src/util/Log.java
@@ -1,4 +1,4 @@
-package Util;
+package util;
 
 import model.Person;
